unify BDFDualSplitting and BDFPressureCorrection regarding momentum/Helmholtz step
