+++ include/exadg/incompressible_navier_stokes/spatial_discretization/operator_dual_splitting.cpp
@@ -57,166 +57,6 @@ OperatorDualSplitting<dim, Number>::~OperatorDualSplitting()
 
 template<int dim, typename Number>
 void
-OperatorDualSplitting<dim, Number>::setup_preconditioners_and_solvers()
-{
-  ProjectionBase::setup_preconditioners_and_solvers();
-
-  setup_helmholtz_preconditioner();
-  setup_helmholtz_solver();
-}
-
-template<int dim, typename Number>
-void
-OperatorDualSplitting<dim, Number>::setup_helmholtz_preconditioner()
-{
-  if(this->param.preconditioner_viscous == PreconditionerViscous::None)
-  {
-    // do nothing, preconditioner will not be used
-  }
-  else if(this->param.preconditioner_viscous == PreconditionerViscous::InverseMassMatrix)
-  {
-    InverseMassOperatorData inverse_mass_operator_data;
-    inverse_mass_operator_data.dof_index  = this->get_dof_index_velocity();
-    inverse_mass_operator_data.quad_index = this->get_quad_index_velocity_linear();
-    inverse_mass_operator_data.parameters = this->param.inverse_mass_preconditioner;
-
-    helmholtz_preconditioner =
-      std::make_shared<InverseMassPreconditioner<dim, dim, Number>>(this->get_matrix_free(),
-                                                                    inverse_mass_operator_data);
-  }
-  else if(this->param.preconditioner_viscous == PreconditionerViscous::PointJacobi)
-  {
-    helmholtz_preconditioner =
-      std::make_shared<JacobiPreconditioner<MomentumOperator<dim, Number>>>(this->momentum_operator,
-                                                                            false);
-  }
-  else if(this->param.preconditioner_viscous == PreconditionerViscous::BlockJacobi)
-  {
-    helmholtz_preconditioner =
-      std::make_shared<BlockJacobiPreconditioner<MomentumOperator<dim, Number>>>(
-        this->momentum_operator, false);
-  }
-  else if(this->param.preconditioner_viscous == PreconditionerViscous::Multigrid)
-  {
-    typedef MultigridPreconditioner<dim, Number> Multigrid;
-
-    helmholtz_preconditioner = std::make_shared<Multigrid>(this->mpi_comm);
-
-    std::shared_ptr<Multigrid> mg_preconditioner =
-      std::dynamic_pointer_cast<Multigrid>(helmholtz_preconditioner);
-
-    std::map<dealii::types::boundary_id, std::shared_ptr<dealii::Function<dim>>>
-      dirichlet_boundary_conditions = this->momentum_operator.get_data().bc->dirichlet_bc;
-
-    // We also need to add DirichletCached boundary conditions. From the
-    // perspective of multigrid, there is no difference between standard
-    // and cached Dirichlet BCs. Since multigrid does not need information
-    // about inhomogeneous boundary data, we simply fill the map with
-    // dealii::Functions::ZeroFunction for DirichletCached BCs.
-    for(auto iter : this->momentum_operator.get_data().bc->dirichlet_cached_bc)
-    {
-      typedef typename std::pair<dealii::types::boundary_id, std::shared_ptr<dealii::Function<dim>>>
-        pair;
-
-      dirichlet_boundary_conditions.insert(
-        pair(iter, new dealii::Functions::ZeroFunction<dim>(dim)));
-    }
-
-    typedef std::map<dealii::types::boundary_id, dealii::ComponentMask> Map_DBC_ComponentMask;
-    Map_DBC_ComponentMask                                               dirichlet_bc_component_mask;
-
-    mg_preconditioner->initialize(this->param.multigrid_data_viscous,
-                                  this->grid,
-                                  this->multigrid_mappings,
-                                  this->get_dof_handler_u().get_fe(),
-                                  this->momentum_operator,
-                                  MultigridOperatorType::ReactionDiffusion,
-                                  this->param.ale_formulation,
-                                  dirichlet_boundary_conditions,
-                                  dirichlet_bc_component_mask);
-  }
-  else
-  {
-    AssertThrow(
-      false, dealii::ExcMessage("Preconditioner specified for viscous step is not implemented."));
-  }
-}
-
-template<int dim, typename Number>
-void
-OperatorDualSplitting<dim, Number>::setup_helmholtz_solver()
-{
-  if(this->param.solver_viscous == SolverViscous::CG)
-  {
-    // setup solver data
-    Krylov::SolverDataCG solver_data;
-    solver_data.max_iter             = this->param.solver_data_viscous.max_iter;
-    solver_data.solver_tolerance_abs = this->param.solver_data_viscous.abs_tol;
-    solver_data.solver_tolerance_rel = this->param.solver_data_viscous.rel_tol;
-
-    if(this->param.preconditioner_viscous == PreconditionerViscous::PointJacobi or
-       this->param.preconditioner_viscous == PreconditionerViscous::BlockJacobi or
-       this->param.preconditioner_viscous == PreconditionerViscous::InverseMassMatrix or
-       this->param.preconditioner_viscous == PreconditionerViscous::Multigrid)
-    {
-      solver_data.use_preconditioner = true;
-    }
-
-    helmholtz_solver = std::make_shared<
-      Krylov::SolverCG<MomentumOperator<dim, Number>, PreconditionerBase<Number>, VectorType>>(
-      this->momentum_operator, *helmholtz_preconditioner, solver_data);
-  }
-  else if(this->param.solver_viscous == SolverViscous::GMRES)
-  {
-    // setup solver data
-    Krylov::SolverDataGMRES solver_data;
-    solver_data.max_iter             = this->param.solver_data_viscous.max_iter;
-    solver_data.solver_tolerance_abs = this->param.solver_data_viscous.abs_tol;
-    solver_data.solver_tolerance_rel = this->param.solver_data_viscous.rel_tol;
-    solver_data.max_n_tmp_vectors    = this->param.solver_data_viscous.max_krylov_size;
-    // use default value of compute_eigenvalues
-
-    // default value of use_preconditioner = false
-    if(this->param.preconditioner_viscous == PreconditionerViscous::PointJacobi or
-       this->param.preconditioner_viscous == PreconditionerViscous::BlockJacobi or
-       this->param.preconditioner_viscous == PreconditionerViscous::InverseMassMatrix or
-       this->param.preconditioner_viscous == PreconditionerViscous::Multigrid)
-    {
-      solver_data.use_preconditioner = true;
-    }
-
-    helmholtz_solver = std::make_shared<
-      Krylov::SolverGMRES<MomentumOperator<dim, Number>, PreconditionerBase<Number>, VectorType>>(
-      this->momentum_operator, *helmholtz_preconditioner, solver_data, this->mpi_comm);
-  }
-  else if(this->param.solver_viscous == SolverViscous::FGMRES)
-  {
-    Krylov::SolverDataFGMRES solver_data;
-    solver_data.max_iter             = this->param.solver_data_viscous.max_iter;
-    solver_data.solver_tolerance_abs = this->param.solver_data_viscous.abs_tol;
-    solver_data.solver_tolerance_rel = this->param.solver_data_viscous.rel_tol;
-    solver_data.max_n_tmp_vectors    = this->param.solver_data_viscous.max_krylov_size;
-
-    if(this->param.preconditioner_viscous == PreconditionerViscous::PointJacobi or
-       this->param.preconditioner_viscous == PreconditionerViscous::BlockJacobi or
-       this->param.preconditioner_viscous == PreconditionerViscous::InverseMassMatrix or
-       this->param.preconditioner_viscous == PreconditionerViscous::Multigrid)
-    {
-      solver_data.use_preconditioner = true;
-    }
-
-    helmholtz_solver = std::make_shared<
-      Krylov::SolverFGMRES<MomentumOperator<dim, Number>, PreconditionerBase<Number>, VectorType>>(
-      this->momentum_operator, *helmholtz_preconditioner, solver_data);
-  }
-  else
-  {
-    AssertThrow(false, dealii::ExcMessage("Specified viscous solver is not implemented."));
-  }
-}
-
-template<int dim, typename Number>
-void
 OperatorDualSplitting<dim, Number>::apply_velocity_divergence_term(VectorType &       dst,
                                                                    VectorType const & src) const
 {
@@ -747,9 +587,13 @@ OperatorDualSplitting<dim, Number>::solve_viscous(VectorType &       dst,
   // Update operator
   this->momentum_operator.set_scaling_factor_mass_operator(factor);
 
-  helmholtz_solver->update_preconditioner(update_preconditioner);
+  // Note that there is no need to set the evaluation time for the momentum_operator
+  // because this function is only called if the convective term is not considered
+  // in the momentum_operator (Stokes eq. or explicit treatment of convective term).
 
-  unsigned int n_iter = helmholtz_solver->solve(dst, src);
+  this->momentum_linear_solver->update_preconditioner(update_preconditioner);
+
+  unsigned int n_iter = this->momentum_linear_solver->solve(dst, src);
 
   return n_iter;
 }
