+++ include/exadg/incompressible_navier_stokes/spatial_discretization/operator_projection_methods.cpp
@@ -80,6 +80,9 @@ OperatorProjectionMethods<dim, Number>::setup_preconditioners_and_solvers()
   setup_solver_pressure_poisson();
 
   Base::setup_projection_solver();
+
+  setup_momentum_preconditioner();
+  setup_momentum_solver();
 }
 
 template<int dim, typename Number>
@@ -252,6 +255,154 @@ OperatorProjectionMethods<dim, Number>::setup_solver_pressure_poisson()
     AssertThrow(false,
                 dealii::ExcMessage(
                   "Specified solver for pressure Poisson equation is not implemented."));
+  }
+}
+
+template<int dim, typename Number>
+void
+OperatorProjectionMethods<dim, Number>::setup_momentum_preconditioner()
+{
+  if(this->param.preconditioner_momentum == MomentumPreconditioner::InverseMassMatrix)
+  {
+    InverseMassOperatorData inverse_mass_operator_data;
+    inverse_mass_operator_data.dof_index  = this->get_dof_index_velocity();
+    inverse_mass_operator_data.quad_index = this->get_quad_index_velocity_linear();
+    inverse_mass_operator_data.parameters = this->param.inverse_mass_preconditioner;
+
+    momentum_preconditioner =
+      std::make_shared<InverseMassPreconditioner<dim, dim, Number>>(this->get_matrix_free(),
+                                                                    inverse_mass_operator_data);
+  }
+  else if(this->param.preconditioner_momentum == MomentumPreconditioner::PointJacobi)
+  {
+    momentum_preconditioner =
+      std::make_shared<JacobiPreconditioner<MomentumOperator<dim, Number>>>(this->momentum_operator,
+                                                                            false);
+  }
+  else if(this->param.preconditioner_momentum == MomentumPreconditioner::BlockJacobi)
+  {
+    momentum_preconditioner =
+      std::make_shared<BlockJacobiPreconditioner<MomentumOperator<dim, Number>>>(
+        this->momentum_operator, false);
+  }
+  else if(this->param.preconditioner_momentum == MomentumPreconditioner::Multigrid)
+  {
+    typedef MultigridPreconditioner<dim, Number> Multigrid;
+
+    momentum_preconditioner = std::make_shared<Multigrid>(this->mpi_comm);
+
+    std::shared_ptr<Multigrid> mg_preconditioner =
+      std::dynamic_pointer_cast<Multigrid>(momentum_preconditioner);
+
+    std::map<dealii::types::boundary_id, std::shared_ptr<dealii::Function<dim>>>
+      dirichlet_boundary_conditions = this->momentum_operator.get_data().bc->dirichlet_bc;
+
+    // We also need to add DirichletCached boundary conditions. From the
+    // perspective of multigrid, there is no difference between standard
+    // and cached Dirichlet BCs. Since multigrid does not need information
+    // about inhomogeneous boundary data, we simply fill the map with
+    // dealii::Functions::ZeroFunction for DirichletCached BCs.
+    for(auto iter : this->momentum_operator.get_data().bc->dirichlet_cached_bc)
+    {
+      typedef typename std::pair<dealii::types::boundary_id, std::shared_ptr<dealii::Function<dim>>>
+        pair;
+
+      dirichlet_boundary_conditions.insert(
+        pair(iter, new dealii::Functions::ZeroFunction<dim>(dim)));
+    }
+
+    typedef std::map<dealii::types::boundary_id, dealii::ComponentMask> Map_DBC_ComponentMask;
+    Map_DBC_ComponentMask                                               dirichlet_bc_component_mask;
+
+    mg_preconditioner->initialize(this->param.multigrid_data_momentum,
+                                  this->grid,
+                                  this->multigrid_mappings,
+                                  this->get_dof_handler_u().get_fe(),
+                                  this->momentum_operator,
+                                  this->param.multigrid_operator_type_momentum,
+                                  this->param.ale_formulation,
+                                  dirichlet_boundary_conditions,
+                                  dirichlet_bc_component_mask);
+  }
+  else
+  {
+    AssertThrow(this->param.preconditioner_momentum == MomentumPreconditioner::None,
+                dealii::ExcNotImplemented());
+  }
+}
+
+template<int dim, typename Number>
+void
+OperatorProjectionMethods<dim, Number>::setup_momentum_solver()
+{
+  if(this->param.solver_momentum == SolverMomentum::CG)
+  {
+    // setup solver data
+    Krylov::SolverDataCG solver_data;
+    solver_data.max_iter             = this->param.solver_data_momentum.max_iter;
+    solver_data.solver_tolerance_abs = this->param.solver_data_momentum.abs_tol;
+    solver_data.solver_tolerance_rel = this->param.solver_data_momentum.rel_tol;
+    if(this->param.preconditioner_momentum != MomentumPreconditioner::None)
+      solver_data.use_preconditioner = true;
+
+    // setup solver
+    momentum_linear_solver = std::make_shared<
+      Krylov::SolverCG<MomentumOperator<dim, Number>, PreconditionerBase<Number>, VectorType>>(
+      this->momentum_operator, *momentum_preconditioner, solver_data);
+  }
+  else if(this->param.solver_momentum == SolverMomentum::GMRES)
+  {
+    // setup solver data
+    Krylov::SolverDataGMRES solver_data;
+    solver_data.max_iter             = this->param.solver_data_momentum.max_iter;
+    solver_data.solver_tolerance_abs = this->param.solver_data_momentum.abs_tol;
+    solver_data.solver_tolerance_rel = this->param.solver_data_momentum.rel_tol;
+    solver_data.max_n_tmp_vectors    = this->param.solver_data_momentum.max_krylov_size;
+    solver_data.compute_eigenvalues  = false;
+    if(this->param.preconditioner_momentum != MomentumPreconditioner::None)
+      solver_data.use_preconditioner = true;
+
+    // setup solver
+    momentum_linear_solver = std::make_shared<
+      Krylov::SolverGMRES<MomentumOperator<dim, Number>, PreconditionerBase<Number>, VectorType>>(
+      this->momentum_operator, *momentum_preconditioner, solver_data, this->mpi_comm);
+  }
+  else if(this->param.solver_momentum == SolverMomentum::FGMRES)
+  {
+    Krylov::SolverDataFGMRES solver_data;
+    solver_data.max_iter             = this->param.solver_data_momentum.max_iter;
+    solver_data.solver_tolerance_abs = this->param.solver_data_momentum.abs_tol;
+    solver_data.solver_tolerance_rel = this->param.solver_data_momentum.rel_tol;
+    solver_data.max_n_tmp_vectors    = this->param.solver_data_momentum.max_krylov_size;
+    if(this->param.preconditioner_momentum != MomentumPreconditioner::None)
+      solver_data.use_preconditioner = true;
+
+    momentum_linear_solver = std::make_shared<
+      Krylov::SolverFGMRES<MomentumOperator<dim, Number>, PreconditionerBase<Number>, VectorType>>(
+      this->momentum_operator, *momentum_preconditioner, solver_data);
+  }
+  else
+  {
+    AssertThrow(false,
+                dealii::ExcMessage("Specified solver for momentum equation is not implemented."));
+  }
+
+
+  // Navier-Stokes equations with an implicit treatment of the convective term
+  if(this->param.nonlinear_problem_has_to_be_solved())
+  {
+    // nonlinear_operator
+    nonlinear_operator.initialize(*this);
+
+    // setup Newton solver
+    momentum_newton_solver = std::make_shared<Newton::Solver<VectorType,
+                                                             NonlinearMomentumOperator<dim, Number>,
+                                                             MomentumOperator<dim, Number>,
+                                                             Krylov::SolverBase<VectorType>>>(
+      this->param.newton_solver_data_momentum,
+      nonlinear_operator,
+      this->momentum_operator,
+      *momentum_linear_solver);
   }
 }
